requirements.txt
# Azure DevOps Integration Agent requirements

# Azure SDK
azure-devops==7.1.0b3
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.16.0
azure-monitor-opentelemetry==1.0.0

# Git interaction
GitPython==3.1.32

# Security
keyring==24.2.0
cryptography==41.0.3

# HTTP and API
requests==2.31.0
httpx==0.24.1

# Command line interface
click==8.1.7
rich==13.5.2

# Parsing and data manipulation
pyyaml==6.0.1
jinja2==3.1.2
beautifulsoup4==4.12.2
lxml==4.9.3
pydantic==2.3.0

# Code parsing and analysis
tree-sitter==0.20.1
astroid==3.0.0
pylint==3.0.1

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Logging and monitoring
structlog==23.1.0
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
colorama==0.4.6